requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
djangorestframework==3.9.2
idna==2.8
pkg-resources==0.0.0
pytz==2019.1
requests==2.21.0
spotipy==2.4.4
sqlparse==0.3.0
urllib3==1.24.2
